+++ temba/settings_engage.py
@@ -413,10 +413,6 @@ DEFAULT_BRAND_OBJ.update({
 
 MSG_FIELD_SIZE = env('MSG_FIELD_SIZE', 4096)
 
-# unset BWI key, causes exception if set and we no longer support it anyway
-BWI_KEY = None
-# above didn't seem to work, so set ENV var to "" for now, debug later.
-
 # set of ISO-639-3 codes of languages to allow in addition to all ISO-639-1 languages
 if env('NON_ISO6391_LANGUAGES_ALLOWED', None) is not None:
     NON_ISO6391_LANGUAGES = tuple(filter(None, env('NON_ISO6391_LANGUAGES_ALLOWED', None).split(',')))
